chore: Build handleSendInitialSms helper
