+++ helpers/index.js
@@ -1,0 +1,3 @@
+import handleSendInitialSms from './handleSendInitialSms';
+
+export { handleSendInitialSms };
