+++ server.js
@@ -4,6 +4,8 @@ const express = require('express');
 const session = require('express-session');
 const Twilio = require('twilio');
 
+const { handleSendInitialSms } = require('./helpers');
+
 // Static variables
 const RESET_KEYWORD = 'reset';
 const EXCLUDED_PROPERTIES = [
@@ -62,39 +64,6 @@ const accessControlAllowOrigin = (request, response, next) => {
     'Origin, X-Requested-With, Content-Type, Accept'
   );
   next();
-};
-
-const handleInitialMessage = async (request, response, twiml) => {
-  try {
-    const continents = await database('continents').select();
-    const continentsList = continents.reduce(
-      (accumulator, { id, name }) => (accumulator += `\t${id} - ${name}\n`),
-      ''
-    );
-
-    twiml.message(`
-      Thanks for writing to Our Planet! 🌎\n\nWe aim to shed light on endangered animals using knowledge from the World Wildlife Foundation alongside beautiful photos 🐻\n\nWho knows, maybe you'll find your new spirit animal! ✨
-    `);
-    twiml.message(
-      `Let's start by selecting a region. Reply with a number and we'll send you a list of endangered animals to learn more about:\n\n${continentsList}`
-    );
-    twiml.message(
-      `Oh! And so you know, we don't store any of your personal information. This is just about the animals 🦉`
-    );
-
-    request.session.animalResponses = DEFAULT_ANIMAL_RESPONSES;
-    response.writeHead(200, { 'Content-Type': 'text/xml' });
-    response.end(twiml.toString());
-  } catch (error) {
-    console.error(error);
-
-    twiml.message(
-      "Whoops! 🍌\n\nLooks like that didn't work. Please try again in a few minutes."
-    );
-
-    response.writeHead(200, { 'Content-Type': 'text/xml' });
-    return response.end(twiml.toString());
-  }
 };
 
 const handleReturnAnimals = async (request, response, twiml) => {
@@ -205,7 +174,7 @@ const handleReturnAnimalById = async (request, response, twiml) => {
         return {
           baseStats: (accumulator.baseStats += `${propertyString}\n\n`),
           facts: accumulator.facts,
-          importance: accumulator.importance
+          importance: accumulator.importance,
         };
       },
       {
@@ -298,7 +267,7 @@ const server = app
 
     // If no session => Create initial, send instruction
     if (!animalResponses || body.Body.toLowerCase() === RESET_KEYWORD) {
-      return handleInitialMessage(request, response, twiml);
+      return handleSendInitialSms(request, response, twiml);
     }
 
     const { countryId, animalId } = animalResponses;
